teste

+++ src/controller/task.js
@@ -4,24 +4,33 @@ class TaskController {
   async criarTask(titulo, descricao, idProjeto, idUsuario) {
     this.validaCampos(titulo, descricao);
     ProjectController.buscarProjeto(idProjeto, idUsuario);
-    return await Task.create({
+    const task = await Task.create({
       titulo,
       descricao,
       id_projeto: idProjeto,
     });
+    return task;
   }
 
-  async alterarTask(id, titulo, descricao, idUsuario, status, dataConclusao) {
+  async alterarTask(id, titulo, descricao, idProjeto, idUsuario, status, dataConclusao) {
     this.validaCampos(titulo, descricao);
-    this.buscarTask(id, idUsuario);
-    return await Task.update(
-      { titulo, descricao, status, dataConclusao },
-      { where: { id } }
-    );
+    ProjectController.buscarProjeto(idProjeto, idUsuario);
+    const task = await this.buscarTask(id, idProjeto);
+    task.titulo = titulo;
+    task.descricao = descricao;
+    task.status = status ? status : task.status;
+    task.data_conclusao = dataConclusao;
+    await task.save();
+    return task;
   }
 
   async deletarTask(id, idUsuario) {
-    this.buscarTask(id, idUsuario);
+    
+    const task = Task.findOne({ where: { id } });
+    if (!task) {
+      throw new Error("Task não encontrada");
+    }
+    ProjectController.buscarProjeto(task.id_projeto, idUsuario);
     await Task.destroy({ where: { id } });
   }
 
@@ -39,11 +48,11 @@ class TaskController {
     });
   }
 
-  async buscarTask(id, idUsuario) {
-    if (idUsuario === undefined || id === undefined) {
+  async buscarTask(id, idProjeto) {
+    if (idProjeto === undefined || id === undefined) {
       throw new Error("Id do usuário e da task são obrigatórios");
     }
-    const task = await Task.findOne({ where: { id, id_usuario: idUsuario } });
+    const task = await Task.findOne({ where: { id, id_projeto: idProjeto } });
     if (!task) {
       throw new Error("Task não encontrado");
     }
